feat(react-18-toast): start refactoring toast package
- storybook works
- some tests are failing
