+++ packages/provider/package.json
@@ -19,6 +19,7 @@
     "@chakra-ui/portal": "1.3.10",
     "@chakra-ui/react-env": "1.1.6",
     "@chakra-ui/system": "1.12.1",
+    "@chakra-ui/toast": "1.5.8",
     "@chakra-ui/utils": "1.10.4"
   },
   "peerDependencies": {
