+++ packages/provider/src/chakra-provider.tsx
@@ -14,6 +14,7 @@ import {
 } from "@chakra-ui/react-env"
 import * as React from "react"
 import { IdProvider } from "@chakra-ui/hooks"
+import { ToastProvider } from "@chakra-ui/toast"
 
 export interface ChakraProviderProps
   extends Pick<ThemeProviderProps, "cssVarsRoot"> {
@@ -74,7 +75,7 @@ export const ChakraProvider: React.FC<ChakraProviderProps> = (props) => {
 
   const _children = (
     <EnvironmentProvider environment={environment}>
-      {children}
+      <ToastProvider>{children}</ToastProvider>
     </EnvironmentProvider>
   )
