+++ packages/toast/package.json
@@ -36,6 +36,8 @@
     "@chakra-ui/alert": "1.3.7",
     "@chakra-ui/close-button": "1.2.7",
     "@chakra-ui/hooks": "1.9.1",
+    "@chakra-ui/portal": "1.3.9",
+    "@chakra-ui/react-utils": "1.2.3",
     "@chakra-ui/theme": "1.14.1",
     "@chakra-ui/transition": "1.4.8",
     "@chakra-ui/utils": "1.10.4",
